requirements.txt
# Core dependencies for clary_quest project
# AI/ML and API dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
pyarrow>=12.0.0

# Network analysis and visualization
networkx>=3.0.0
matplotlib>=3.7.0

# Web requests (for deprecated modules)
requests>=2.31.0

# Optional: Better graph layouts (requires graphviz system package)
# pygraphviz>=1.10
